+++ src/App.js
@@ -1,10 +1,15 @@
 import Home from "./pages/Home";
 import "./App.css";
+import { RouterProvider, createBrowserRouter } from "react-router-dom";
+import { routes } from "./routes/routes";
+const router = createBrowserRouter(routes);
 function App() {
   return (
-    <div className="App">
-      <Home />
-    </div>
+    <RouterProvider router={router}>
+      <div className="App">
+        <Home />
+      </div>
+    </RouterProvider>
   );
 }
