+++ src/components/rightbar/Rightbar.jsx
@@ -3,7 +3,6 @@ import React, { useState } from "react";
 import Banner from "../banner/Banner";
 import SearchFilter from "../searchFilter/SearchFilter";
 import Playlist from "../playlist/Playlist";
-import Filters from "../fitlers/Filters";
 
 function Rightbar() {
   const [showFilter, setShowFilter] = useState(false);
@@ -12,14 +11,13 @@ function Rightbar() {
     <Box sx={{ flex: 18 }}>
       <Banner />
       <Box sx={{ margin: "0 138px" }}>
-        <SearchFilter showFilter={showFilter} setShowFilter={setShowFilter} />
-        {console.log(filterVariable)}
-        {showFilter && (
-          <Filters
-            filterVariable={filterVariable}
-            setFilterVariable={setFilterVariable}
-          />
-        )}
+        <SearchFilter
+          showFilter={showFilter}
+          setShowFilter={setShowFilter}
+          filterVariable={filterVariable}
+          setFilterVariable={setFilterVariable}
+        />
+
         <Playlist filterVariable={filterVariable} />
       </Box>
     </Box>
